server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
humanize==4.9.0
python-multipart==0.0.6
Pillow
opencv-python-headless
aiofiles
